Pubsub_Kafka/Code/Subscriber/requirements.txt
kafka-python==2.0.2
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
docopt==0.6.2
Flask==2.0.1
Flask-RESTful==0.3.9
grip==4.5.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
path-and-address==2.0.1
Pygments==2.10.0
pysolr==3.9.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tweepy==4.0.0
urllib3==1.26.7
Werkzeug==2.0.1
